# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Social-Media-Api
  version: v1
  description: "Спецификация для проекта Social-Media-Api."
servers:
  - url: http://localhost:8080
    description: local server
paths:
...
  /api/v1/chat/get:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Chat endpoints
      summary: "Получение чата по id"
      operationId: getChat
      parameters:
        - name: chatId
          description: "ID чата"
          in: query
          required: true
          allowEmptyValue: false
          example: 1
          schema:
            type: integer
      responses:
        '200':
          description: Successful get chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatDto'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/chat/last_message:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Chat endpoints
      summary: "Получение последнего сообщения в чате по id чата"
      operationId: getLastMessageFromChat
      parameters:
        - name: chatId
          description: "ID чата"
          in: query
          required: true
          allowEmptyValue: false
          example: 1
          schema:
            type: integer
      responses:
        '200':
          description: Successful get last message at chat
          content:
            '*/*':
              schema:
                type: string
                example: "Последнее сообщение в чате"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/chat/create:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Chat endpoints
      summary: "Создание чата"
      operationId: createChat
      requestBody:
        description: "Ввод id первого и второго пользователей для создания чата между ними"
        content:
          application/json:
            schema:
              properties:
                chatDto:
                  $ref: '#/components/schemas/ChatDto'
      responses:
        '201':
          description: Successful create chat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/chat/join:
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Chat endpoints
      summary: "Присоединение к чату"
      operationId: joinChat
      parameters:
        - name: chatId
          description: "ID чата"
          in: query
          required: true
          allowEmptyValue: false
          example: 1
          schema:
            type: integer
      responses:
        '202':
          description: Successful join into chat
          content:
            '*/*':
              schema:
                type: string
                example: "User with id = 1 join to chat with id = 1"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/chat/leave:
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Chat endpoints
      summary: "Отключение от чата"
      operationId: leaveChat
      parameters:
        - name: chatId
          description: "ID чата"
          in: query
          required: true
          allowEmptyValue: false
          example: 1
          schema:
            type: integer
      responses:
        '202':
          description: Successful join into chat
          content:
            '*/*':
              schema:
                type: string
                example: "User with id = 1 leave to chat with id = 1"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/chat/delete:
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - Chat endpoints
      summary: "Удаление чата по id"
      operationId: deleteChat
      parameters:
        - name: chatId
          description: "ID чата"
          in: query
          required: true
          allowEmptyValue: false
          example: 1
          schema:
            type: integer
      responses:
        '200':
          description: Successful delete chat
          content:
            '*/*':
              schema:
                type: string
                example: "Chat with id = 1 deleted!"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/getAllChats:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Chat endpoints
      summary: "Получение всех существующих чатов между пользователями (Доступен только авторизованным пользователям с ролью ADMIN)"
      operationId: getAllChats
      responses:
        '200':
          description: Successful get all chats from BD
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatDto'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChatDto:
      required:
        - firstUserId
        - secondUserId
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: "Id чата"
          example: "1"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: "Дата и время создания чата"
          example: "2024-01-12 13:56"
          pattern: "yyyy-MM-dd HH:mm"
        lastMessage:
          type: string
          readOnly: true
          description: "Последенее сообщение в чате"
          example: "Последенее сообщение в чате"
        firstUserId:
          type: integer
          description: "Id первого пользователя в чате"
          example: "1"
        secondUserId:
          type: integer
          description: "Id второго пользователя в чате"
          example: "2"
...
    ErrorResponse:
      type: object
      properties:
        messageError:
          type: string
          description: Error message
          example: "Error"